Refactor the demo component into a JobStatusPoller that predominantly just polls for updates and re-renders the canvas, and a wrapper that determines all of the metadata and node fetching methods.
